Add the function to generate a new program for the parallel call
This function generates only one program based on the input
parameters. Can be called iteratively and, therefore, parallely
to populate the population of programs.
This also evaluate the fitness for the first time and, being
parallelizable, it is quite convenient. The only drawback is the
need of quite a lot input parameters to pass also the fitness
functions and their parameters (data, target and weights when
required).

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>