Improves order_preserving introducing mse computation
The order_preserving fitness function allow to evaluate the cardinal number
of violations in ordering datapoint with respect to a given target value.
Now it also include the magnitude of the error and support the return of
the number of inversions only, the magnitude only or both combined in a
"double" fitness.
All the fitness evaluations in the codebase are adapted accordingly to support
an eventual double returned value.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>