Introduces ANN-based constants optimization
A program is constituted by a set of operations and operands.
Operands can be features chosen from the training dataset and
numeric constants. The genetic evolution of the programs is based
on a random process and there might be a situation in which the
overall structure of a formula is correct for the learning task,
but the constants within the formula are not the optimal one.
Waiting for those constants to mutate and become a better version
of themselves to improve performance is a non-optimal strategy
and it can be improved leveraging a gradient based optimization
to alter the constants and give them the best value possible
given the current formula structure and the current objective
functions.
To achieve this we introduce a ANN-based approach which uses
a 1-neuron neural network in which the activation function is
the program itself and the weights in input to that neuron are
the constants of the formula. The gradiendt descent algorithm
modifies the weights and bring them to a value that produces
better performance.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>