Refactor the modelling of the fitness functions and code structure
Fitness functions were once simple functions passed to the trainer
through a dictionary and were evaluated by just accessing the
element of the dictionary corresponding to the callable. It came
also with additional configuration like the convergence threshold
and the minimization specification.
Now all fitness are reimplemented as classes in which we can set
all the configuration parameters and execute them to access the
performance metric.

This required a complete restructuring of the whole algorithm
to adapt to the new format.

Finally, this refactoring brought inside the SymbolicRegressor
class many of the functions that were once implemented in other
files to make the repository leaner and cleaner.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>