Add the tree diversity calculation of the population
The diversity is the degree to which two individuals of the
populations are different from each other based on the number
of common sub-trees of the expression. The hashing of a program
is a list of all subtrees in an inverse Polish notation.
Comparing the hasing of two programs we can identify the ratio
of sub-trees they have in common. The tree diversity of a
population is 1- the average diversity of all individuals of the
first pareto front with each other.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>