Remove data storage in SymbolicRegressor to comply with FedML
Previously, the Symbolic Regressor saved the input data for the
training in self.data, but this is not compliant with the idea
of Federated Learning by which no data is shared.
Because the SR object is shared in the federated process, here
we remove the saving of the data from this object and therefore
pass data as argument every time it is needed without storing it.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>