Introduce population as an Object that is serialized for optimization
Now the population is an extension of Python lists that allow the conversion
to binary object of individual programs. Additionally, the domination lists
that are used for the NSGA-II algorithm are emptied when not needed so to
prevent DeepRecursionError both in the parallelization dispatch and the saving
of the population to the checkpoint file.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>