Add FedAvgNSGAII Federated aggregation strategy
This aggregation strategy execute in two stages for each federated
round. This needed BaseStrategy to incorporate the concept of
federated_round_is_terminated to control the increment of the
number of rounds executed. Only when the second stage is completed
the round is considered terminated and can move to the following.

The stages are described as follows for clients and servers.
- Phase 1 Client: train a MOSR population
- Phase 1 Server: collect the 1PF and 2PF from all clients'
poulations
- Phase 2 Client: receive the combined PFs from the server and
optimize the constants on the local dataset
- Phase 2 Server: calculate the constant-wise average of each
program and generate a new population in which each program has
the average of the constants of the programs in the same position
of the previous populations.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>