Add feature confidence interval to FeatureNode
A feature node can be assigned a constant value or a
variable. When we assign a constant value, we can decide
to assign a set of possible values (e.g., from a KFold
experiment) and generate confidence intervals.
The final value will be the median of the provided ones,
whereas the confidence intervals will be a two-elements
list with the 5th and 95th percentiles of the provided
values.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>