Fix first_pareto_front_history missing deepcopy
At the end of each generation the first pareto front is
appended into a history that allow to go back in time
for better analysis of the training process.
Without a deepcopy, if a program was present in more
than one PF, every time it was modified, the modification
was propagated also in the history part which should
be a frozen shapshot instead.
Now a deepcopy of the PF is created preventing this
from happening.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>