Add callbacks support in the fit method.
A number of method are executed at precise points in the
training. Namely, on_initialization_start and end,
on_generation_start and end, on_offspring_generation_start
and end, on_refill_start and end, on_stats_calculation_start
and end, on_convergence and on_training_completed.

The user will define a Callback object extending the
MOSRCallbackBase and implement the methods desired.
To access the regressor from within the callback and
get or set attributes, use the attribute callback.sr
that is set for each of them by the setter inside the
regressor implementation.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>