Removes timeout in offspring generation
The timeout was implemented at a stage when the optimization
of the execution was not yet achieved and allowed to kill
processes that were stuck in the generation of offsprings.
This caused incompatibilities between different OS and
caused Windows machines to struggle to execute in parallel.
Now with no timeout required, this has been removed and
the execution does not have problems anymore.

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>