Add support to fitness functions with export
Exporting the output of a fitness function is used to prevent
recomputation when not needed. The fitness functions should then
be developed to support it.
When exporting, the return value of the .evaluate() method of
the function is a tuple where the first element is the usual
return value (e.g., the WMSE); the second element is a dictionary
in which the key are the label of the exported variable and the
value, its value. The following functions will be passed an
argument "inject" that is a dictionary containing those values
and they can use them to prevent their recomputation

Signed-off-by: Davide Ferrari <davide.ferrari@kcl.ac.uk>